feat(accounts): add preferred_name field to UserProfileSerializer
This change adds the `preferred_name` field to the
`UserProfileSerializer`. This allows the API to return the
preferred name of the user in the user profile.

feat(workflow): add status map for purchase orders

This change adds a status map for purchase orders. This map is
used to transform the Xero purchase order status to our internal
PurchaseOrder model status.